[UVM] initial support of UVM in Kaapi with dynamic selection. added functions: kaapi_offload_properties, kaapi_offload_get_uvm_capacity, kaapi_offload_set_force_uvm To use UVM capability with Kaapi : - configure library with ENABLE_KAAPI_UNIFIED. Default is OFF. Once compiled and installed, you can run your program with the env. var KAAPI_FORCED_UVM defined as: KAAPI_FORCED_UVM=0 -> do not force to use UVM, this is the default KAAPI_FORCED_UVM=1 -> if ENABLE_KAAPI_UNIFIED is ON, and the hardware are basic capability to use UVM then force mode is active and the runtime do not send explicit data through cudaMemcpyXYAsync function calls. To check the UVM state: XKBLAS_VERBOSE=1 with display additional line:   UVM      : yes - unused or   UVM      : yes - used where yes correspond to the hardware capability of the library to support UVM (iff ENABLE_KAAPI_UNIFIED is on). used or unused correspond to if uvm was used.

+++ blas/xkblas.c
@@ -2899,7 +2899,7 @@ void xkblas_free_work_pos( void* ptr )
 		{
 			break;
 		}
-		if((uintptr_t) curr->ptr > ptr)
+		if((uintptr_t) curr->ptr > (uintptr_t)ptr)
 		{
 			curr = NULL;
 			break;
@@ -3014,6 +3014,9 @@ struct memory_buffer_args {
 	size_t element_size;
 };
 
+
+/*
+*/
 void* memset_init_thread_function(void* args)
 {
 	struct memory_buffer_args* bargs = (struct memory_buffer_args*) args;
@@ -3040,6 +3043,10 @@ void* memset_init_thread_function(void* args)
 	free( args );
 	return NULL;
 }	
+
+
+/*
+*/
 void xkblas_register_work_buffer( void* ptr, size_t size )
 {
 	pthread_mutex_lock( &work_buffer_mutex );
@@ -3073,7 +3080,7 @@ void xkblas_register_work_buffer( void* ptr, size_t size )
 	printf("advise ok\n");
 	
 	struct memory_buffer_args *bargs = (struct memory_buffer_args*) malloc(sizeof(struct memory_buffer_args));
-	bargs->pos =  (uintptr_t) ptr;
+	bargs->pos =  ptr;
 	bargs->size = size;
 	bargs->element_size = 1024L * 1024L * 1024L;
